+++ drivers/iommu/qcom/Makefile
@@ -3,5 +3,4 @@ obj-$(CONFIG_QCOM_IOMMU_V1) += qcom-iommu.o
 qcom-iommu-y += msm_iommu.o
 qcom-iommu-y += msm_iommu-v1.o
 qcom-iommu-y += msm_iommu_dev-v1.o
-qcom-iommu-y += msm_iommu_sec.o
 qcom-iommu-y += msm_iommu_pagetable.o
